requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
boto3==1.20.37
botocore==1.23.37
certifi==2022.9.24
charset-normalizer==2.1.1
cohesion==1.0.0
colorama==0.4.4
Django==3.2.10
django-ses==2.3.1
django-session-timeout==0.1.0
dnspython==2.2.1
email-validator==1.3.0
future==0.18.2
idna==3.4
jmespath==0.10.0
mando==0.6.4
mssql-django==1.1.1
Pillow==9.0.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
radon==5.1.0
reportlab==3.6.2
requests==2.28.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
